# storage/scribe/openapi.yaml
openapi: 3.0.3
info:
  title: 'Evolução local de snipe-it API Documentation'
  description: 'This documentation aims to provide all the information you need to work with the Snipe-IT API.'
  version: 1.0.0
servers:
  -
    url: 'https://snipe-it.test'
...
paths:
...
  '/api/v1/categories/{item_type}/selectlist':
    get:
      summary: Selectlist
      operationId: selectlist
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Unauthorized or unauthenticated.'
                properties:
                  error:
                    type: string
                    example: 'Unauthorized or unauthenticated.'
      tags:
        - Categories
    parameters:
      -
        in: path
        name: item_type
        description: ''
        example: architecto
        required: true
        schema:
          type: string
...
  /api/v1/manufacturers:
    get:
      summary: 'List Manufacturers'
      operationId: listManufacturers
      description: ''
      parameters:
        -
          in: query
          name: search
          description: 'Search term to filter results.'
          example: Dell
          required: false
          schema:
            type: string
            description: 'Search term to filter results.'
            example: Dell
            nullable: false
        -
          in: query
          name: name
          description: 'Filter by exact manufacturer name.'
          example: Dell
          required: false
          schema:
            type: string
            description: 'Filter by exact manufacturer name.'
            example: Dell
            nullable: false
        -
          in: query
          name: url
          description: 'Filter by exact URL.'
          example: 'http://example.com'
          required: false
          schema:
            type: string
            description: 'Filter by exact URL.'
            example: 'http://example.com'
            nullable: false
        -
          in: query
          name: support_url
          description: 'Filter by exact support URL.'
          example: 'http://support.example.com'
          required: false
          schema:
            type: string
            description: 'Filter by exact support URL.'
            example: 'http://support.example.com'
            nullable: false
        -
          in: query
          name: warranty_lookup_url
          description: 'Filter by exact warranty lookup URL.'
          example: 'http://warranty.example.com'
          required: false
          schema:
            type: string
            description: 'Filter by exact warranty lookup URL.'
            example: 'http://warranty.example.com'
            nullable: false
        -
          in: query
          name: support_phone
          description: 'Filter by exact support phone number.'
          example: 1-800-555-5555
          required: false
          schema:
            type: string
            description: 'Filter by exact support phone number.'
            example: 1-800-555-5555
            nullable: false
        -
          in: query
          name: support_email
          description: 'Filter by exact support email address.'
          example: support@example.org
          required: false
          schema:
            type: string
            description: 'Filter by exact support email address.'
            example: support@example.org
            nullable: false
        -
          in: query
          name: sort
          description: 'Column to sort results by. Allowed values: id, name, url, support_url, support_email, warranty_lookup_url, support_phone, created_at, updated_at, assets_count, consumables_count, components_count, licenses_count. Default: created_at.'
          example: name
          required: false
          schema:
            type: string
            description: 'Column to sort results by. Allowed values: id, name, url, support_url, support_email, warranty_lookup_url, support_phone, created_at, updated_at, assets_count, consumables_count, components_count, licenses_count. Default: created_at.'
            example: name
            nullable: false
        -
          in: query
          name: order
          description: 'Order of sorted results. Allowed values: asc, desc. Default: desc.'
          example: asc
          required: false
          schema:
            type: string
            description: 'Order of sorted results. Allowed values: asc, desc. Default: desc.'
            example: asc
            nullable: false
        -
          in: query
          name: offset
          description: 'Offset/starting position of the results. Default: 0.'
          example: 0
          required: false
          schema:
            type: integer
            description: 'Offset/starting position of the results. Default: 0.'
            example: 0
            nullable: false
        -
          in: query
          name: limit
          description: 'Limit the number of results returned. Default: 25. Maximum: 100.'
          example: 50
          required: false
          schema:
            type: integer
            description: 'Limit the number of results returned. Default: 25. Maximum: 100.'
            example: 50
            nullable: false
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Unauthorized or unauthenticated.'
                properties:
                  error:
                    type: string
                    example: 'Unauthorized or unauthenticated.'
      tags:
        - Manufacturers
    post:
      summary: 'Create Maintenance'
      operationId: createMaintenance
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Manufacturers
...